Add files via upload
Generative adversarial network (GANs) is a new breed of artificial intelligence technology introduced by Ian Goodfellow in 2014 for creating fake media that looks real. This was popularly known as deepfake.
GANs are an approach to generative models that use neural networks to produce synthetic data via a game of two adversaries: The Generator and the Discriminator. 
The goal of the generator is to create new and plausible samples that are convincing enough to fool the discriminator.
The discriminator on the other hand takes a sample from the domain as input and predicts a binary class target of real or fake.
The generator and the discriminator motivate each other to improve their capabilities as training continues until a point where the generator produces realistic images, and the discriminator fails at distinguishing fake from real ones.
This research was motivated by astronomers’ need to test algorithms that measure the parameters of galaxies. The study and classification of galaxies in parametric research can be challenging and computationally intensive, hence a set of artificial galaxies become necessary.